backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
openai==1.6.1
anthropic==0.8.1
pydantic==2.5.3
sqlalchemy==2.0.23
redis==5.0.1
pytest==7.4.3
httpx==0.25.2
python-multipart==0.0.6
websockets==12.0
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
aiofiles==23.2.1
jinja2==3.1.2
gitpython==3.1.40
ast-grep-py==0.12.0
faker
sqlalchemy
alembic
psycopg2-binary
